requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-classic
pypdf
chromadb
sentence-transformers
python-dotenv
